fix(gpu-client): fix HWND type for windows-sys 0.59 (isize → *mut c_void)

+++ nexterm-client-gpu/src/renderer.rs
@@ -2013,7 +2013,8 @@ fn apply_acrylic_blur(window: &winit::window::Window) {
 
     let Ok(handle) = window.window_handle() else { return };
     let RawWindowHandle::Win32(h) = handle.as_raw() else { return };
-    let hwnd = h.hwnd.get() as isize;
+    // windows-sys 0.59 では HWND = *mut c_void。as isize キャストは不要
+    let hwnd = h.hwnd.get();
 
     // DWMWA_SYSTEMBACKDROP_TYPE = 38; 4 = DWMWCP_ACRYLIC（Windows 11 22H2+）
     let backdrop_type: u32 = 4;
